application/slidea/requirements.txt
aiofiles
aiosqlite
httpx
langchain
langchain_core
langgraph
langchain_openai
langchain_text_splitters
langgraph_checkpoint_sqlite
loguru
python-dotenv
pydantic_settings
numpy
json_repair
jsonschema
tavily-python
langchain_unstructured
unstructured
unstructured[md,docx,doc,pptx]
PyPDF2
fake_useragent
pdfminer.six==20240706
python-pptx
playwright
llvmlite==0.45.0
